Don't install more executables than necessary
The gemspec is configured to install everything in `bin` as executables.
However, this directory contains not only `armadura`, but `rspec`,
`rake`, and `setup` as well. These do not need to become first-class
executables when Armadura is installed.

In order to fix this, move the `armadura` executable to `exe` and
configure the gemspec to use this directory instead to install
executables.